Big merge from commercial branch ...
GUI:
 - Redshift - greatly enhanced support
 - Connection base class: don't mangle case in sth_2_sqlite()
 - Redshift: Fix LongTruncOk parameter setting
 - 1st run wizard:
  - layout improvements
  - enable ODBCINSTINI management by default
  - end of wizard enhancements - documentation and full restart improvements
 - configuration window:
  - layout and documentation improvements
  - disable debug mode in connections form
  - hide some hacks that the commercial version requires
 - main window:
  - job list/editor layout enhancements - form editor for jobs now
  - param list/editor layout enhancements - form editor for parameters now
  - disable debug mode in a bunch of forms/datasheets
 - window base class:
  - don't print to STDOUT in lots of places ( remove debugging )
 - framework launcher - lots of escaping fixes when building cli args
 - gui_logs ( STDOUT / STDERR ) improvements:
  - filter GLib-GObject-WARNING STDERR lines - this is bogus and noisy

ETL Runtime:
 - Redshift support
 - AWS S3 support:
  - S3_GET
  - S3_PUT
  - S3_ITERATOR
 - DB name is no longer required when generating a database class
 - Postgres:
  - fixes for subclasses
  - don't mangle case of keys in recordsets except for METADATA connections
 - Add trailing slash in #ENV_JOB_LOG_DIR#
 - CompressFile:
  - remove double slashes in paths
  - fix flatpak paths for compression/decompression binaries
 - DecompressFile - new template
 - DataCop: add #P_MAX_ISSUES# logic

Metadata and other assets::
  - S3 templates
  - DataCop template
  - CompressFile / DecompressFile templates & icons
  - SQLServer BCP tempaltes & icons

+++ smart_etl/SmartAssociates/TemplateConfig/CompressFile.pm
@@ -27,6 +27,8 @@ sub execute {
     my $delete_original  = $self->resolve_parameter( '#P_DELETE_ORIGINAL#' );
     
     my $template_text    = $self->detokenize( $template_config->{TEMPLATE_TEXT} );
+
+    $filename =~ s/\/\//\//g; # remove double slashes - we might need to handle paths "properly"
     
     eval {
         
@@ -35,7 +37,7 @@ sub execute {
         if ( $compression_type eq 'gzip' ) {
             
             # First check for pigz - a parallel compression utility
-            $app_path = "/usr/bin/pigz";
+            $app_path = "/app/bin/pigz";
             
             if ( -f $app_path ) {
                 
@@ -53,7 +55,7 @@ sub execute {
                 
             } else {
                 
-                $app_path = "/bin/gzip";
+                $app_path = "/usr/bin/gzip";
                                 
                 if ( -f $app_path ) {
                     
@@ -77,7 +79,7 @@ sub execute {
             
         } elsif ( $compression_type eq 'bzip2' ) {
             
-            $app_path = "/bin/bzip2";
+            $app_path = "/usr/bin/bzip2";
             
             if ( -f $app_path ) {
                 
@@ -126,7 +128,7 @@ sub execute {
             
         }
         
-        if ( $delete_original ) {
+        if ( $delete_original && -e $filename ) {
             
             unlink $filename
                 or die( "Could not unlink $filename: $!" );
